git-svn-id: http://theory.cm.utexas.edu/svn/eon@505 9d962951-508e-48a8-94b5-974f9cfda47c

+++ client/potentials/EAM/EAM.cpp
@@ -39,7 +39,7 @@ void EAM::cleanMemory()
 
 
 // Calculate here long num_cells, long *num_axis, long *cell_length, //become global variables -long *celllist_old, long *celllist_new, long *neigh_list, long fcalled)
-void EAM::force(long N, const double *R, const long *atomicNrs, double *F, double *U, const double *box)
+void EAM::force(long N, const double *R, const int *atomicNrs, double *F, double *U, const double *box)
 {
 
     /* -- Code starting here only needs to be done once. -- */
@@ -175,7 +175,7 @@ EAM::element_parameters EAM::get_element_parameters(int atomic_number)
 }
 
 
-void EAM::calc_force(long N, double *R, const long *atomicNrs, double *F, double *U, const double *box)
+void EAM::calc_force(long N, double *R, const int *atomicNrs, double *F, double *U, const double *box)
 {
     double *drho_dr = new double[3*N]; 
     for (long i=0;i<3*N;i++)
