+++ client/potentials/EMT/EffectiveMediumTheory.h
@@ -31,6 +31,6 @@ public:
     
     // To satify interface
     void initialize() {};
-    void force(long N, const double *R, const long *atomicNrs, double *F, double *U, const double *box);
+    void force(long N, const double *R, const int *atomicNrs, double *F, double *U, const double *box);
 };
 #endif
